% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsriver.R
\name{newsriver_key}
\alias{newsriver_key}
\title{Setup}
\usage{
newsriver_key(key)
}
\arguments{
\item{key}{Your API key, freely available at \url{https://newsriver.io}.}
}
\description{
Setup API key to be used for subsequent calls.
}
\note{
You can specify \code{NEWSRIVER_API_KEY} as environment variable, 
likely in your \code{.Renviron} file.
}
\examples{
\dontrun{
newsriver_key("xXXxxXxXxXxx")
}

}
