% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsriver.R
\name{nr_search}
\alias{nr_search}
\title{Search}
\usage{
nr_search(q, sort_by = NULL, sort_order = NULL, results = 100)
}
\arguments{
\item{q}{Your search query, i.e.: \code{text:(Barack AND Obama)}, more details \url{https://console.newsriver.io/code-book}.}

\item{sort_by}{How should the search result be sorted. By default, articles are sorted by date in descending order (newest items first).}

\item{sort_order}{Sort order can be either \code{ASC} (ascending) or \code{DESC} (descending). The sort order is ignored if the search is sorted by (\code{sort_by}) \code{_score}.}

\item{results}{Number of results to return}
}
\description{
Search for articles.
}
\section{Sort}{

Articles can be sorted by: 
\itemize{
  \item{\code{_score} - the search matching score (most relevant items first).}
  \item{\code{discoverDate} - When the article has been first found online (typically close to publication date).}
  \item{\code{metadata.readTime.seconds} - Article's reading time.}  
}
Other metadata fields can be used for sorting, but search performance 
may be affected as only the fields mentioned above are currently indexed for sorting.
}

\examples{
\dontrun{
newsriver_key("xXXxxXxXxXxx")
articles <- nr_search('language:en AND title:"New York"')
}

}
